requirements.txt
Pillow
datetime
streamlit
ultralytics
